# mall-admin/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: mall admin
  version: 1.0.0
  title: Mall admin
basePath: /admin
schemes:
  - http

#
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#
paths:
...
  /sys/user/list:
    get:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: page
          description: 
          in: query
          type: integer
          required: true
        - name: limit
          description: 
          in: query
          type: integer
          required: true
        - name: sidx
          description: 
          in: query
          type: string
        - name: order
          description: ascdesc
          in: query
          type: string
        - name: username
          description: 
          in: query
          type: string
      responses:
        '200':
          description: 
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: userId
          description: ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 
          schema:
            type: object
            properties:
              code:
                description:  0  0
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: body
          description: 
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: body
          description: 
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: body
          description: 
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 
      summary: 
      produces:
        - application/json
      parameters:
        - name: body
          description: ID
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description:  0  0
          type: integer
          format: int32
        msg:
          description: 
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 
        type: string
      newPassword:
        description: 
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: ID
        type: integer
        format: int64
      username:
        description: 
        type: string
      password:
        description: 
        type: string
      email:
        description: 
        type: string
      mobile:
        description: 
        type: string
      status:
        description:   0   1
        type: integer
        format: int32
      roleIdList:
        description: ID
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: ID
        type: integer
        format: int64
      createTime:
        description: 
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description:  0  0
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 
            type: integer
            format: int32
          pageSize:
            description: 
            type: integer
            format: int32
          totalPage:
            description: 
            type: integer
            format: int32
          currPage:
            description: 
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: ID
          type: integer
          format: int64
        username:
          description: 
          type: string
        password:
          description: 
          type: string
        email:
          description: 
          type: string
        mobile:
          description: 
          type: string
        status:
          description:   0   1
          type: integer
          format: int32
        roleIdList:
          description: ID
          type: array
          items:
            type: integer
            format: int32
...